Fix input/output for py3k.
In py3k, decoding/encoding of text files is done automatically during
read/write operations and the write() method fails with encoded (i.e.
binary) strings. For performance, and universal newline support, we
use the built-in encoding mechanism instead of the "homegrown" one if
py3k is detected; as a consequence, io.FileIO.write() returns a
Unicode string (the correct argument type for file.write() in py3k).